refactor: Differentiating AttackingEventArgs
