+++ EXILED/Exiled.Events/EventArgs/Scp106/Scp106AttackingEventArgs.cs
@@ -5,41 +5,41 @@
 // </copyright>
 // -----------------------------------------------------------------------
 
-namespace Exiled.Events.EventArgs.Scp049
+namespace Exiled.Events.EventArgs.Scp106
 {
     using Exiled.API.Features;
     using Exiled.API.Features.Roles;
     using Exiled.Events.EventArgs.Interfaces;
 
     /// <summary>
-    /// Contains all information before SCP-049 attacks a player.
+    /// Contains all information before SCP-106 attacks a player.
     /// </summary>
-    public class AttackingEventArgs : IScp049Event, IDeniableEvent
+    public class Scp106AttackingEventArgs : IScp106Event, IDeniableEvent
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="AttackingEventArgs"/> class.
+        /// Initializes a new instance of the <see cref="Scp106AttackingEventArgs"/> class.
         /// </summary>
         /// <param name="player"><inheritdoc cref="Player"/></param>
         /// <param name="target"><inheritdoc cref="Target"/></param>
         /// <param name="isAllowed"><inheritdoc cref="IsAllowed"/></param>
-        public AttackingEventArgs(Player player, Player target, bool isAllowed = true)
+        public Scp106AttackingEventArgs(Player player, Player target, bool isAllowed = true)
         {
             Player = player;
-            Scp049 = player.Role.As<Scp049Role>();
+            Scp106 = player.Role.As<Scp106Role>();
             Target = target;
             IsAllowed = isAllowed;
         }
 
-        /// <inheritdoc/>
-        public Scp049Role Scp049 { get; }
-
         /// <summary>
-        /// Gets the player controlling SCP-049.
+        /// Gets the player controlling SCP-106.
         /// </summary>
         public Player Player { get; }
 
+        /// <inheritdoc/>
+        public Scp106Role Scp106 { get; }
+
         /// <summary>
-        /// Gets the target of attack.
+        /// Gets the target of the attack.
         /// </summary>
         public Player Target { get; }
